refactor(FormGroup): wrapperRef宣言・useDescribedByIds・ref統合をFormGroupへ集約する
FormControl/Fieldsetに重複していたwrapperRefの宣言、useDescribedByIds
の呼び出し、useMergeRefsによるwrapperRef+autoBindErrorCallbackRefの
統合をFormGroup側にまとめた。FormControl/Fieldsetはコンポーネント固有
のcallbackRefのみをFormGroupへ渡すようにし、FormGroupのwrapperRef prop
はcallbackRefにリネームした。

Fieldset固有だったwrapper要素へのaria-describedby付与は、FormGroup側
でas === 'fieldset'により判定するようにした。FormGroupはuseRef/
useMergeRefsを直接使うようになったため'use client'を追加し、逆に
Fieldsetはuserefを使わなくなったため'use client'を削除した(rollup
ビルド出力とreact-server条件での評価で動作確認済み)。
